+++ src/client/app/App.jsx
@@ -1,6 +1,7 @@
 /* @flow */
 import React, { PropTypes } from 'react';
-import { ConnectedRouter } from 'connected-react-router/immutable';
+import { Style } from 'radium';
+import { ConnectedRouter } from 'connected-react-router';
 
 import routes from './routes';
 
@@ -9,9 +10,23 @@ import routes from './routes';
  */
 function App({ history }) {
   return (
-    <ConnectedRouter history={history}>
-      { routes }
-    </ConnectedRouter>
+    <div>
+      <Style
+        rules={{
+          '*': {
+            boxSizing: 'border-box',
+          },
+          body: {
+            background: '#eee',
+            color: '#111',
+            fontFamily: '"Roboto", sans-serif',
+          },
+        }}
+      />
+      <ConnectedRouter history={history}>
+        { routes }
+      </ConnectedRouter>
+    </div>
   );
 }