+++ src/client/app/components/PageNotFound/PageNotFound.jsx
@@ -1,19 +1,24 @@
 /* @flow */
 import React from 'react';
 import Helmet from 'react-helmet';
+import Radium from 'radium';
 
-import styles from './PageNotFound.scss';
+const styles = {
+  header: {
+    fontSize: '20px',
+  },
+};
 
 /**
  * Page not found component.
  */
 function PageNotFound() {
   return (
-    <div className={styles.pageNotFound}>
+    <div>
       <Helmet title="404" />
-      <h1>404</h1>
+      <h1 style={[styles.header]}>404</h1>
     </div>
   );
 }
 
-export default PageNotFound;
+export default Radium(PageNotFound);