+++ src/client/app/components/Todo/Todo.jsx
@@ -1,17 +1,26 @@
 /* @flow */
 import React, { PropTypes } from 'react';
+import Radium from 'radium';
 
-import styles from './Todo.scss';
+const styles = {
+  todo: {
+    fontSize: '18px',
+    userSelect: 'none',
+  },
+  completed: {
+    textDecoration: 'line-through',
+  },
+};
 
 function Todo({ onClick, completed, text }) {
   return (
     // eslint-disable-next-line jsx-a11y/no-static-element-interactions
     <li
-      className={styles.todo}
+      style={[
+        styles.todo,
+        completed && styles.completed,
+      ]}
       onClick={onClick}
-      style={{
-        textDecoration: completed ? 'line-through' : 'none',
-      }}
     >
       {text}
     </li>
@@ -24,4 +33,4 @@ Todo.propTypes = {
   text: PropTypes.string.isRequired,
 };
 
-export default Todo;
+export default Radium(Todo);