+++ src/client/app/components/Todo/Todo.spec.jsx
@@ -8,10 +8,10 @@ describe('Todo', () => {
   it('renders without crashing', () => {
     const todo = mount(
       <Todo
-        onClick={function() { }}
+        onClick={function noop() { }}
         text={'Todo'}
         completed={false}
-      />
+      />,
     );
 
     expect(todo).toMatchSnapshot();
@@ -21,10 +21,10 @@ describe('Todo', () => {
     it('displays todo text', () => {
       const todo = shallow(
         <Todo
-          onClick={function() { }}
+          onClick={function noop() { }}
           text={'Todo'}
           completed={false}
-        />
+        />,
       );
 
       const expected = 'Todo';
@@ -36,31 +36,25 @@ describe('Todo', () => {
     it('displays strikethrough if completed', () => {
       const todo = shallow(
         <Todo
-          onClick={function() { }}
+          onClick={function noop() { }}
           text={'Todo'}
-          completed={true}
-        />
+          completed
+        />,
       );
 
-      const expected = 'line-through';
-      const actual = todo.props().style.textDecoration;
-
-      expect(actual).toBe(expected);
+      expect(todo.find('li').props().style.textDecoration).toEqual('line-through');
     });
 
     it('displays no strikethrough if incomplete', () => {
       const todo = shallow(
         <Todo
-          onClick={function() { }}
+          onClick={function noop() { }}
           text={'Todo'}
           completed={false}
-        />
+        />,
       );
 
-      const expected = 'none';
-      const actual = todo.props().style.textDecoration;
-
-      expect(actual).toBe(expected);
+      expect(todo.find('li').props().style.textDecoration).toEqual(undefined);
     });
 
     it('fires onclick when clicked', () => {
@@ -71,24 +65,12 @@ describe('Todo', () => {
           onClick={mockFn}
           text={'Todo'}
           completed={false}
-        />
+        />,
       );
 
       todo.find('li').simulate('click');
 
       expect(mockFn).toHaveBeenCalledTimes(1);
-    });
-
-    it('has todo class', () => {
-      const todo = shallow(
-        <Todo
-          onClick={function() { }}
-          text={'Todo'}
-          completed={false}
-        />
-      );
-
-      expect(todo.find('li').hasClass('todo')).toBe(true);
     });
   });
 });