+++ src/client/app/components/TodoList/TodoList.jsx
@@ -1,12 +1,18 @@
 /* @flow */
 import React, { PropTypes } from 'react';
+import Radium from 'radium';
 
-import Todo from '../Todo/Todo';
-import styles from './TodoList.scss';
+import Todo from '../Todo';
+
+const styles = {
+  todoList: {
+    cursor: 'pointer',
+  },
+};
 
 function TodoList({ todos, onTodoClick }) {
   return (
-    <ul className={styles.todoList}>
+    <ul style={[styles.todoList]}>
       {todos.map(todo =>
         <Todo
           key={todo.id}
@@ -19,12 +25,14 @@ function TodoList({ todos, onTodoClick }) {
 }
 
 TodoList.propTypes = {
-  todos: PropTypes.arrayOf(PropTypes.shape({
-    id: PropTypes.number.isRequired,
-    completed: PropTypes.bool.isRequired,
-    text: PropTypes.string.isRequired,
-  }).isRequired).isRequired,
+  todos: PropTypes.arrayOf(
+    PropTypes.shape({
+      id: PropTypes.number.isRequired,
+      completed: PropTypes.bool.isRequired,
+      text: PropTypes.string.isRequired,
+    }).isRequired,
+  ).isRequired,
   onTodoClick: PropTypes.func.isRequired,
 };
 
-export default TodoList;
+export default Radium(TodoList);