+++ src/client/app/containers/TodoList.jsx
@@ -2,10 +2,11 @@
 import { connect } from 'react-redux';
 
 import { toggleTodo } from '../actions';
-import TodoList from '../components/TodoList/TodoList';
+import { getTodos } from '../selectors';
+import TodoList from '../components/TodoList';
 
 const mapStateToProps = state => ({
-  todos: state.get('todos').toJS(),
+  todos: getTodos(state),
 });
 
 const mapDispatchToProps = dispatch => ({