+++ src/client/main.hmr.jsx
@@ -4,22 +4,19 @@ import { render } from 'react-dom';
 import { createStore, applyMiddleware, compose } from 'redux';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
-import * as Immutable from 'immutable';
 import { createBrowserHistory } from 'history';
-import { connectRouter, routerMiddleware } from 'connected-react-router/immutable';
+import { connectRouter, routerMiddleware } from 'connected-react-router';
 import { IntlProvider } from 'react-intl-redux';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import injectTapEventPlugin from 'react-tap-event-plugin';
+import { StyleRoot } from 'radium';
 
 import { AppContainer } from 'react-hot-loader';
 import createLogger from 'redux-logger';
 
-import App from './app/App';
+import App from './app';
 import rootReducer from './app/reducers';
 import DevTools from './app/containers/DevTools';
-
-// global styles
-import './main.scss';
 
 // favicon
 import './favicon.ico';
@@ -27,13 +24,7 @@ import './favicon.ico';
 injectTapEventPlugin();
 
 const history = createBrowserHistory();
-const initialState = Immutable.fromJS({
-  intl: {
-    locale: 'en',
-    messages: {},
-  },
-});
-const intlSelector = state => state.get('intl').toJS();
+const initialState = {};
 
 const store = createStore(
   connectRouter(history)(rootReducer),
@@ -42,9 +33,7 @@ const store = createStore(
     applyMiddleware(
       routerMiddleware(history),
       thunk,
-      createLogger({
-        stateTransformer: state => state.toJS(),
-      }),
+      createLogger(),
     ),
     DevTools.instrument(),
   ),
@@ -54,10 +43,12 @@ const renderApp = () => {
   render(
     <AppContainer>
       <Provider store={store}>
-        <IntlProvider intlSelector={intlSelector}>
+        <IntlProvider>
           <MuiThemeProvider>
             <div>
-              <App history={history} />
+              <StyleRoot>
+                <App history={history} />
+              </StyleRoot>
               <DevTools />
             </div>
           </MuiThemeProvider>
@@ -71,7 +62,7 @@ const renderApp = () => {
 renderApp();
 
 if (module.hot) {
-  module.hot.accept('./app/App', renderApp);
+  module.hot.accept('./app', renderApp);
 
   module.hot.accept('./app/reducers', () => {
     store.replaceReducer(connectRouter(history)(rootReducer));