+++ webpack/webpack.common.js
@@ -1,6 +1,5 @@
 const webpack = require('webpack');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-const SassLintPlugin = require('sasslint-webpack-plugin');
 
 module.exports = {
   resolve: {
@@ -22,7 +21,7 @@ module.exports = {
       },
       {
         test: /\.(png|jpe?g|gif|svg)$/,
-        use: 'file-loader?name=assets/images/[name].[hash].[ext]'
+        use: 'file-loader?name=assets/img/[name].[hash].[ext]'
       },
       {
         test: /\.ico$/,
@@ -32,19 +31,15 @@ module.exports = {
   },
 
   plugins: [
-    new SassLintPlugin({
-      glob: 'src/client/**/*.s?(a|c)ss'
-    }),
     new webpack.optimize.CommonsChunkPlugin({
-      name: ['app', 'vendor', 'polyfills']
+      name: ['app', 'vendor', 'polyfill']
     }),
     new HtmlWebpackPlugin({
       template: 'src/client/index.html'
     }),
     new webpack.DefinePlugin({
       'process.env': {
-        'NODE_ENV': JSON.stringify(process.env.NODE_ENV),
-        'APP_NAME': JSON.stringify(process.env.APP_NAME)
+        NODE_ENV: JSON.stringify(process.env.NODE_ENV)
       }
     })
   ],