+++ webpack/webpack.dev.js
@@ -2,15 +2,15 @@ const path = require('path');
 const webpack = require('webpack');
 const Visualizer = require('webpack-visualizer-plugin');
 
-const port = 8080;
+const { PORT, HOST, PROXY_PORT } = process.env;
 
 module.exports = {
   entry: {
-    polyfills: './src/client/polyfills',
+    polyfill: './src/client/polyfill',
     vendor: './src/client/vendor',
     app: [
       'react-hot-loader/patch',
-      `webpack-dev-server/client?http://localhost:${port}`,
+      `webpack-dev-server/client?http://${HOST}:${PORT}`,
       'webpack/hot/only-dev-server',
       './src/client/main.hmr'
     ]
@@ -21,22 +21,8 @@ module.exports = {
   output: {
     path: path.resolve(__dirname, '../dist'),
     publicPath: '/',
-    filename: 'assets/[name].js',
-    chunkFilename: 'assets/[id].chunk.js'
-  },
-
-  module: {
-    rules: [
-      {
-        test: /\.s?(a|c)ss$/,
-        use: [
-          'style-loader',
-          'css-loader?modules&importLoaders=2&camelCase',
-          'postcss-loader',
-          'sass-loader'
-        ]
-      }
-    ]
+    filename: 'assets/js/[name].js',
+    chunkFilename: 'assets/js/[id].chunk.js'
   },
 
   plugins: [
@@ -54,9 +40,10 @@ module.exports = {
     compress: true,
     contentBase: path.resolve(__dirname, '../dist'),
     publicPath: '/',
-    port,
+    port: PORT,
+    host: HOST,
     proxy: {
-      '/api': 'http://localhost:3000'
+      '/api': `http://${HOST}:${PROXY_PORT}`
     }
   }
 };