+++ webpack/webpack.prod.js
@@ -1,11 +1,10 @@
 const path = require('path');
 const webpack = require('webpack');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const Visualizer = require('webpack-visualizer-plugin');
 
 module.exports = {
   entry: {
-    polyfills: './src/client/polyfills',
+    polyfill: './src/client/polyfill',
     vendor: './src/client/vendor',
     app: './src/client/main'
   },
@@ -15,20 +14,8 @@ module.exports = {
   output: {
     path: path.resolve(__dirname, '../dist'),
     publicPath: '/',
-    filename: 'assets/[name].[hash].js',
-    chunkFilename: 'assets/[id].[hash].chunk.js'
-  },
-
-  module: {
-    rules: [
-      {
-        test: /\.s?(a|c)ss$/,
-        loader: ExtractTextPlugin.extract({
-          fallbackLoader: 'style-loader',
-          loader: 'css-loader?importLoaders=2&minimize&modules&camelCase!postcss-loader!sass-loader'
-        })
-      }
-    ]
+    filename: 'assets/js/[name].[hash].js',
+    chunkFilename: 'assets/js/[id].[hash].chunk.js'
   },
 
   plugins: [
@@ -38,9 +25,6 @@ module.exports = {
       compress: {
         warnings: false
       }
-    }),
-    new ExtractTextPlugin({
-      filename: 'assets/[name].[hash].css'
     }),
     new Visualizer({
       filename: '../webpack/stats/stats.prod.html'